# openapi.yaml
...
openapi: 3.1.0
servers:
  - url: http://127.0.0.1:4021
info:
  title: Narrative Engine API
  summary: Narrative creation and management service for story conception and structured content.
  description: |
    This service provides tools for creating and managing narrative projects. It is designed primarily for
    authors to conceive, structure, and edit their stories using configurable modules and schemas.
    The service is driven by AI-assistance for the creative process.

    ## Modules

    Modules define the structure for narrative content. Each module includes a JSON Schema that describes
    the expected data format, along with UI configuration for rendering interactive editors. Modules provide
    a framework for organizing story elements—whether filled in manually by authors or generated with AI assistance.

    ## Projects

    Projects are user-owned containers that group related schemas together. Each project has a workflow that
    defines which modules are available for content creation.

    ## Schemas

    Schemas are the content instances created within a project. They conform to a module's structure and
    serve as the building blocks for story conception. Authors can create and edit schemas manually through
    the UI, or optionally use AI generation as a starting point or creative aid.

    ## Authentication

    This service requires an authenticated token, that can be retrieved from the 
...
  version: v0.0.1
  license:
    name: AGPL-3.0
...
paths:
...
  /schemas:
    get:
      operationId: schemaSelect
      summary: Retrieve a specific schema.
      description: |
        Retrieve a schema by its ID or by project and module combination.
        The user must own the project containing the schema. The projectID is always required
        for ownership verification. When retrieving by module, both projectID and module are needed.
      tags: [schemas]
      security:
        - BearerAuth: ["schemas:get"]
      parameters:
        - $ref: "#/components/parameters/schemaID"
        - $ref: "#/components/parameters/projectID"
        - name: module
          in: query
          description: The module identifier in `namespace:id@vX.X.X` or `namespace:id@vX.X.X-preversion` format. Required when not providing a schema ID.
          required: false
          schema:
            type: string
            examples: ["agora:idea@v1.0.0"]
      responses:
        "200":
          $ref: "#/components/responses/schemaSelect"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/notFound"
        "422":
          $ref: "#/components/responses/unprocessableEntity"
        default:
          $ref: "#/components/responses/internalError"
    put:
      operationId: schemaCreate
      summary: Create a new schema.
      description: |
        Create a new schema within a project for manual authoring or to save content from any source.
        The user must own the project and the module must be part of the project's workflow.
      tags: [schemas]
      security:
        - BearerAuth: ["schemas:create"]
      requestBody:
        $ref: "#/components/requestBodies/schemaCreate"
      responses:
        "201":
          $ref: "#/components/responses/schemaSelect"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/notFound"
        "409":
          $ref: "#/components/responses/conflict"
        "422":
          $ref: "#/components/responses/unprocessableEntity"
        default:
          $ref: "#/components/responses/internalError"
    patch:
      operationId: schemaRewrite
      summary: Update schema data.
      description: |
        Update the data content of an existing schema, typically through manual editing in the UI.
        The user must own the project containing the schema.
      tags: [schemas]
      security:
        - BearerAuth: ["schemas:rewrite"]
      requestBody:
        $ref: "#/components/requestBodies/schemaRewrite"
      responses:
        "200":
          $ref: "#/components/responses/schemaSelect"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/notFound"
        "422":
          $ref: "#/components/responses/unprocessableEntity"
        default:
          $ref: "#/components/responses/internalError"
  /schemas/versions:
    get:
      operationId: schemaListVersions
      summary: List schema versions.
      description: |
        Retrieve a paginated list of schema versions for a specific project and module.
        The user must own the project.
      tags: [schemas]
      security:
        - BearerAuth: ["schemas:versions:list"]
      parameters:
        - $ref: "#/components/parameters/projectID"
        - $ref: "#/components/parameters/moduleIDRequired"
        - $ref: "#/components/parameters/moduleNamespaceRequired"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          $ref: "#/components/responses/schemaListVersions"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/notFound"
        "422":
          $ref: "#/components/responses/unprocessableEntity"
        default:
          $ref: "#/components/responses/internalError"
  /schemas/generate:
    put:
      operationId: schemaGenerate
      summary: Generate a schema using AI assistance.
      description: |
        Optionally generate a new schema using AI based on the module's configuration.
        This endpoint provides AI assistance as an alternative to manual content creation.
        The user must own the project and the module must be part of the project's workflow.
      tags: [schemas]
      security:
        - BearerAuth: ["schemas:generate"]
      requestBody:
        $ref: "#/components/requestBodies/schemaGenerate"
      responses:
        "201":
          $ref: "#/components/responses/schemaSelect"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/notFound"
        "409":
          $ref: "#/components/responses/conflict"
        "422":
          $ref: "#/components/responses/unprocessableEntity"
        default:
          $ref: "#/components/responses/internalError"
components:
  responses:
...
    schemaSelect:
      description: The schema details.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/schema"
    schemaListVersions:
      description: List of schema versions.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/schemaVersion"
    unauthorized:
      description: |
        The request did not include valid authentication credentials.
    forbidden:
      description: |
        The user was authenticated successfully, but its current access rights don't grant permission for
        this operation, or the user does not own the requested resource.
    unprocessableEntity:
      description: |
        The request was understood by the server, but cannot be processed because the data did not pass
        required validation.
    notFound:
      description: |
        The requested data was not found on the server.
    badRequest:
      description: |
        The request sent to the server could not be parsed.
    conflict:
      description: |
        The record already exists.
    internalError:
      description: Something unexpected happened.
...
  schemas:
...
    schema:
      type: object
      description: A content instance conforming to a module's structure. Schemas are typically created and edited manually by authors, with optional AI assistance available.
      required: [id, projectID, module, source, data, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/uuid"
        projectID:
          $ref: "#/components/schemas/uuid"
        owner:
          oneOf:
            - $ref: "#/components/schemas/uuid"
            - type: "null"
          description: The owner's user ID. May be null for system-generated schemas.
        module:
          type: string
          description: The module identifier in `namespace:id@vX.X.X` or `namespace:id@vX.X.X-preversion` format.
          examples: ["agora:idea@v1.0.0"]
        source:
          $ref: "#/components/schemas/schemaSource"
        data:
          type: object
          description: The actual content data conforming to the module's schema.
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the schema was created.
          examples: [2009-11-10T23:00:00Z]
    schemaVersion:
      type: object
      description: A version entry for a schema.
      required: [id, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/uuid"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the version was created.
          examples: [2009-11-10T23:00:00Z]
    uuid:
      type: string
      description: A universally unique identifier.
      format: uuid
      examples:
        - "9dce0fa2-f93b-46a9-aa6b-a71bf0b1ee80"
    moduleIDField:
      type: string
      description: The module identifier.
      examples: ["idea", "character"]
    moduleNamespaceField:
      type: string
      description: The module namespace.
      examples: ["agora", "custom"]
...
    schemaSource:
      type: string
      description: Indicates how the schema content was created (USER for manual authoring, AI for generated content, FORK for copied from another schema, EXTERNAL for imported content).
      enum: ["USER", "AI", "FORK", "EXTERNAL"]
    lang:
      type: string
      format: ISO-639
      description: Language code for the content.
      enum: [en, fr]
    limit:
      type: integer
      format: int32
      minimum: 1
      maximum: 100
      description: Maximum number of results to return.
    offset:
      type: integer
      format: int32
      minimum: 0
      description: Number of results to skip for pagination.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    moduleIDRequired:
      name: moduleID
      in: query
      description: The module ID.
      required: true
      schema:
        $ref: "#/components/schemas/moduleIDField"
    moduleNamespaceRequired:
      name: moduleNamespace
      in: query
      description: The module namespace.
      required: true
      schema:
        $ref: "#/components/schemas/moduleNamespaceField"
    schemaID:
      name: id
      in: query
      description: The schema ID.
      required: false
      schema:
        $ref: "#/components/schemas/uuid"
    projectID:
      name: projectID
      in: query
      description: The project ID.
      required: true
      schema:
        $ref: "#/components/schemas/uuid"
    limit:
      name: limit
      in: query
      description: Limit the number of results.
      required: false
      schema:
        $ref: "#/components/schemas/limit"
    offset:
      name: offset
      in: query
      description: Skip results for pagination.
      required: false
      schema:
        $ref: "#/components/schemas/offset"
  requestBodies:
...
    schemaCreate:
      description: Request to create a new schema.
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [id, projectID, module, source, data]
            properties:
              id:
                $ref: "#/components/schemas/uuid"
              projectID:
                $ref: "#/components/schemas/uuid"
              module:
                type: string
                description: The module identifier in `namespace:id@vX.X.X` or `namespace:id@vX.X.X-preversion` format.
                examples: ["agora:idea@v1.0.0"]
              source:
                $ref: "#/components/schemas/schemaSource"
              data:
                type: object
                description: The content data conforming to the module's schema.
                additionalProperties: true
    schemaRewrite:
      description: Request to update schema data.
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [id, data]
            properties:
              id:
                $ref: "#/components/schemas/uuid"
              data:
                type: object
                description: The updated content data.
                additionalProperties: true
    schemaGenerate:
      description: Request to generate a schema using AI assistance.
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [projectID, module, lang]
            properties:
              projectID:
                $ref: "#/components/schemas/uuid"
              module:
                type: string
                description: The module identifier in `namespace:id@vX.X.X` or `namespace:id@vX.X.X-preversion` format.
                examples: ["agora:idea@v1.0.0"]
              lang:
                $ref: "#/components/schemas/lang"